feat(04): switch to Pusher broadcasting and fix integration issues
- Switch BROADCAST_CONNECTION from reverb to pusher
- Configure Echo with pusher-js import and Pusher constructor
- Add broadcastAs() to EnrollmentStatusChanged for correct event name matching
- Fix MQTT publish topic: mqtt/face/{device_id} (no /Edit suffix)
- Fix ACK handler: read AddSucInfo/AddErrInfo from nested info object
- Fix errcode type cast: camera sends string, translateErrorCode expects int
- Add separate 'publisher' MQTT connection to avoid client ID collision
- Use JSON_UNESCAPED_SLASHES for MQTT payloads
- Update all test mocks for new MQTT and ACK structure

+++ app/Mqtt/Handlers/AckHandler.php
@@ -127,7 +127,7 @@ class AckHandler implements MqttHandler
             }
 
             $errorMessage = app(CameraEnrollmentService::class)
-                ->translateErrorCode($item['errcode'] ?? 0);
+                ->translateErrorCode((int) ($item['errcode'] ?? 0));
 
             $enrollment = CameraEnrollment::where('camera_id', $cameraId)
                 ->whereHas('personnel', fn ($q) => $q->where('custom_id', $customId))
